+++ lib/macinbox/actions/check_macos_versions.rb
@@ -8,7 +8,6 @@ module Macinbox
         @installer_app     = opts[:installer_path]  or raise ArgumentError.new(":installer_path not specified")
 
         @collector         = opts[:collector]       or raise ArgumentError.new(":collector not specified")
-        @debug             = opts[:debug]
 
         raise Macinbox::Error.new("Installer app not found") unless File.exist? @installer_app
       end
@@ -21,13 +20,13 @@ module Macinbox
         installer_os_version_components = installer_os_version.split(".") rescue [0, 0, 0]
         installer_os_version_major = installer_os_version_components[0]
         installer_os_version_minor = installer_os_version_components[1]
-        Logger.info "Installer macOS version detected: #{installer_os_version}" if @debug
+        Logger.info "Installer macOS version detected: #{installer_os_version}" if $verbose
 
         host_os_version = Task.backtick %W[ /usr/bin/sw_vers -productVersion ]
         host_os_version_components = host_os_version.split(".") rescue [0, 0, 0]
         host_os_version_major = host_os_version_components[0]
         host_os_version_minor = host_os_version_components[1]
-        Logger.info "Host macOS version detected: #{host_os_version}" if @debug
+        Logger.info "Host macOS version detected: #{host_os_version}" if $verbose
 
         if installer_os_version_major != host_os_version_major || installer_os_version_minor != host_os_version_minor
           Logger.error "Warning: host OS version (#{host_os_version}) and installer OS version (#{installer_os_version}) do not match"