+++ lib/macinbox/actions/create_box_from_hdd.rb
@@ -25,7 +25,6 @@ module Macinbox
         @hidpi             = opts[:hidpi]
 
         @collector         = opts[:collector]       or raise ArgumentError.new(":collector not specified")
-        @debug             = opts[:debug]
 
         raise Macinbox::Error.new("HDD not found") unless File.exist? @input_hdd
       end
@@ -55,7 +54,7 @@ module Macinbox
             end
           EOF
 
-          task_opts = @debug ? {} : { :out => File::NULL }
+          task_opts = $verbose ? {} : { :out => File::NULL }
 
           Task.run %W[ prlctl create macinbox -o macos --no-hdd --dst #{@box_dir} ] + [task_opts]