+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next version (Unreleased)
+## 3.3.0 (February 26, 2019)
 
 FEATURES: